iox-#707 fixed mac os warning: uint64_t == unsigned long (linux), uint64_t == unsigned long long (mac os)

+++ iceoryx_examples/waitset_in_c/ice_c_waitset_gateway.c
@@ -130,7 +130,7 @@ int main()
             }
         }
 
-        printf("sum of all samples: %lu\n", sumOfAllSamples);
+        printf("sum of all samples: %lu\n", (unsigned long)sumOfAllSamples);
         fflush(stdout);
     }
